# Add basic benchmarks and fix some performance issues