--- a/.travis.yml
+++ b/.travis.yml
@@ -8,9 +8,9 @@ julia:
   - nightly
 notifications:
   email: false
-# matrix:
-#   allow_failures:
-#     - julia: nightly
+matrix:
+  allow_failures:
+    - julia: nightly
 #     - os: osx
 # uncomment the following lines to override the default test script
 #script: