--- a/src/angleaxis_types.jl
+++ b/src/angleaxis_types.jl
@@ -43,9 +43,10 @@ end
     AngleAxis{promote_type(promote_type(promote_type(Θ, X), Y), Z)}(θ, x, y, z, normalize)
 end
 
-# These 2 functions are enough to satisfy the entire StaticArrays interface:
+# These functions are enough to satisfy the entire StaticArrays interface:
 @inline (::Type{AA}){AA <: AngleAxis}(t::NTuple{9}) = AA(Quat(t))
 @inline Base.getindex(aa::AngleAxis, i::Int) = Quat(aa)[i]
+@inline Tuple(aa::AngleAxis) = Tuple(Quat(aa))
 
 @inline function Base.convert{R <: RotMatrix}(::Type{R}, aa::AngleAxis)
     # Rodrigues' rotation formula.
@@ -85,8 +86,6 @@ end
     theta =  2 * atan2(s, q.w)
     return s > 0 ? AA(theta, q.x / s, q.y / s, q.z / s) : AA(theta, one(theta), zero(theta), zero(theta))
 end
-
-@inline Base.convert(::Type{Tuple}, aa::AngleAxis) = Tuple(Quat(aa))
 
 # Using Rodrigues formula on an AngleAxis parameterization (assume unit axis length) to do the rotation
 # (implementation from: https://ceres-solver.googlesource.com/ceres-solver/+/1.10.0/include/ceres/rotation.h)
@@ -147,9 +146,10 @@ end
 # but this isn't quite what we mean when we have 4 inputs (not 9).
 @inline (::Type{RodriguesVec}){X,Y,Z}(x::X, y::Y, z::Z) = RodriguesVec{promote_type(promote_type(X, Y), Z)}(x, y, z)
 
-# These 2 functions are enough to satisfy the entire StaticArrays interface:
+# These functions are enough to satisfy the entire StaticArrays interface:
 @inline (::Type{RV}){RV <: RodriguesVec}(t::NTuple{9}) = RV(Quat(t))
 @inline Base.getindex(aa::RodriguesVec, i::Int) = Quat(aa)[i]
+@inline Tuple(rv::RodriguesVec) = Tuple(Quat(rv))
 
 # define its interaction with other angle representations
 @inline Base.convert{R <: RotMatrix}(::Type{R}, rv::RodriguesVec) = convert(R, AngleAxis(rv))
@@ -180,7 +180,6 @@ function Base.convert{RV <: RodriguesVec}(::Type{RV}, q::Quat)
     return RV(sc * q.x, sc * q.y, sc * q.z )
 end
 
-@inline Base.convert(::Type{Tuple}, rv::RodriguesVec) = Tuple(Quat(rv))
 
 function Base.:*{T1,T2}(rv::RodriguesVec{T1}, v::StaticVector{T2})
     if length(v) != 3