--- a/test/REQUIRE
+++ b/test/REQUIRE
@@ -1,1 +1,2 @@
 ForwardDiff
+BenchmarkTools