Add some basic rotation type conversion benchmarks. Output as of right now:
```
Rotations.SPQuat{Float64} -> Rotations.RodriguesVec{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             859.833 ns
  gctime:           0.000 ns (0.00%)
  memory:           752 bytes
  allocs:           24
Rotations.SPQuat{Float64} -> Rotations.SPQuat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.730 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.SPQuat{Float64} -> Rotations.AngleAxis{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             806.691 ns
  gctime:           0.000 ns (0.00%)
  memory:           624 bytes
  allocs:           15
Rotations.RodriguesVec{Float64} -> Rotations.Quat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             36.798 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.Quat{Float64} -> Rotations.RodriguesVec{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             84.546 ns
  gctime:           0.000 ns (0.00%)
  memory:           144 bytes
  allocs:           9
Rotations.Quat{Float64} -> Rotations.AngleAxis{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             21.316 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.AngleAxis{Float64} -> Rotations.SPQuat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             916.097 ns
  gctime:           0.000 ns (0.00%)
  memory:           608 bytes
  allocs:           15
Rotations.Quat{Float64} -> Rotations.Quat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.783 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.SPQuat{Float64} -> Rotations.Quat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             8.377 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.Quat{Float64} -> Rotations.SPQuat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             2.964 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RotMatrix{3,Float64,9} -> Rotations.Quat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             16.981 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RotMatrix{3,Float64,9} -> Rotations.AngleAxis{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             57.857 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RotMatrix{3,Float64,9} -> Rotations.RotMatrix{3,Float64,9}:
BenchmarkTools.TrialEstimate: 
  time:             2.080 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.AngleAxis{Float64} -> Rotations.RodriguesVec{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.730 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RodriguesVec{Float64} -> Rotations.AngleAxis{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             13.116 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RodriguesVec{Float64} -> Rotations.RotMatrix{3,Float64,9}:
BenchmarkTools.TrialEstimate: 
  time:             31.930 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RodriguesVec{Float64} -> Rotations.SPQuat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.080 μs
  gctime:           0.000 ns (0.00%)
  memory:           608 bytes
  allocs:           15
Rotations.RodriguesVec{Float64} -> Rotations.RodriguesVec{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.450 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.Quat{Float64} -> Rotations.RotMatrix{3,Float64,9}:
BenchmarkTools.TrialEstimate: 
  time:             645.228 ns
  gctime:           0.000 ns (0.00%)
  memory:           656 bytes
  allocs:           15
Rotations.AngleAxis{Float64} -> Rotations.RotMatrix{3,Float64,9}:
BenchmarkTools.TrialEstimate: 
  time:             15.995 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.AngleAxis{Float64} -> Rotations.AngleAxis{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             1.451 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RotMatrix{3,Float64,9} -> Rotations.SPQuat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             22.274 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.SPQuat{Float64} -> Rotations.RotMatrix{3,Float64,9}:
BenchmarkTools.TrialEstimate: 
  time:             791.435 ns
  gctime:           0.000 ns (0.00%)
  memory:           656 bytes
  allocs:           15
Rotations.AngleAxis{Float64} -> Rotations.Quat{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             20.048 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0
Rotations.RotMatrix{3,Float64,9} -> Rotations.RodriguesVec{Float64}:
BenchmarkTools.TrialEstimate: 
  time:             115.546 ns
  gctime:           0.000 ns (0.00%)
  memory:           144 bytes
  allocs:           9
```
Just quickly examining these results, I found:
* `RodriguesVec` and `SPQuat` have the same issue as https://github.com/FugroRoames/Rotations.jl/pull/38
* conversion from `Quat` to `RotMatrix` calls `Tuple` constructor in `StaticArrays` code. `Rotations` defines a `convert` method from `Quat` to `Tuple`, but calling the `Tuple` constructor doesn't call this `convert` method.
* `Quat` -> `RodriguesVec` is type unstable (variable `sc`)