@tkoolen 2017-07-30
Resolved all of the issues reported in the PR description and allowed failures on nightly because StaticArrays does the same.