@andyferris 2017-08-03
Thanks for doing this and fixing the performance issues!